refactor: promote skill profile into jobs
