+++ simaple/data/jobs/__init__.py
@@ -1,0 +1,5 @@
+from simaple.data.jobs._builtin import get_skill_profile
+
+__all__ = [
+    "get_skill_profile",
+]
