+++ simaple/data/jobs/definitions/__init__.py
@@ -1,0 +1,5 @@
+from simaple.data.jobs.definitions.skill_profile import SkillProfile
+
+__all__ = [
+    "SkillProfile",
+]
