+++ simaple/data/jobs/resources/_skill_profile/adele.yaml
