+++ simaple/data/jobs/resources/_skill_profile/archmagefb.yaml
