+++ simaple/data/jobs/resources/_skill_profile/archmagetc.yaml
