+++ simaple/data/jobs/resources/_skill_profile/bishop.yaml
