+++ simaple/data/jobs/resources/_skill_profile/dualblade.yaml
