+++ simaple/data/jobs/resources/_skill_profile/mechanic.yaml
