+++ simaple/data/jobs/resources/_skill_profile/soulmaster.yaml
