+++ simaple/data/jobs/resources/_skill_profile/windbreaker.yaml
