+++ simaple/simulate/interface/simulator_configuration.py
@@ -6,7 +6,8 @@ from simaple.container.simulation import SimulationContainer, SimulationSetting
 from simaple.core.base import ActionStat, Stat
 from simaple.core.jobtype import JobType
 from simaple.data import get_damage_logic
-from simaple.data.skill_profile import SkillProfile, get_skill_profile
+from simaple.data.jobs.definitions import SkillProfile
+from simaple.data.jobs import get_skill_profile
 from simaple.simulate.base import SimulationRuntime
 from simaple.simulate.builder import EngineBuilder
 from simaple.simulate.engine import OperationEngine
